delete table added

+++ src/Customer/ManageCustomer.php
@@ -13,6 +13,11 @@ $result = mysqli_query($conn,"SELECT * FROM details");
 <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
 <link href="https://fonts.googleapis.com/css?family=Poppins:600&display=swap" rel="stylesheet">
 <link rel="stylesheet" href="../../public/css/Manage.css">
+<script type="text/javascript">
+function confirm_alert(node) {
+    return confirm("Do you really want to delete?");
+}
+</script>
 
 </head>
 <body>
@@ -62,8 +67,8 @@ if (mysqli_num_rows($result) > 0) {
 			    <td><?php echo $row["phoneno"]; ?></td>
 			    <td><?php echo $row["email"]; ?></td>
 			    
-				<td><a href="updated.php?id=<?php echo $row["id"]; ?>"><i class="fa fa-pencil" style="font-size:25px;color:blue"></i></a></td>
-				<td><a href="deleted.php?id=<?php echo $row["id"]; ?>"><i class="fa fa-trash" style="font-size:25px;color:red"></i></a></td>
+				<td><a href="updated.php?id=<?php echo $row["id"]; ?> "><i class="fa fa-pencil" style="font-size:25px;color:blue"></i></a></td>
+				<td><a href="deleted.php?id=<?php echo $row["id"]; ?>" onclick="return confirm_alert(this)";><i class="fa fa-trash" style="font-size:25px;color:red"></i></a></td>
 
 			  </tr>
 			  	<?php
